hv: riscv: fix strict-aliasing error in local_irq_save
Fix a compilation error that may occur in some environments due to an
inconsistent type cast. The compiler reports a strict-aliasing violation:
acrn-hypervisor/hypervisor/include/arch/riscv/asm/cpu.h:73:18:
error: dereferencing type-punned pointer will break strict-aliasing rules
[-Werror=strict-aliasing]
   73 |                 *(uint32_t *)(x) = val;
      |                  ^~~~~

Tracked-On: #8813

+++ hypervisor/include/arch/riscv/asm/cpu.h
@@ -132,9 +132,9 @@ static inline void arch_asm_pause(void)
 #define local_irq_restore(x) ({ asm volatile("csrs sstatus, %0\n" ::"rK"(x & SSTATUS_SIE) : "memory"); })
 #define local_irq_save(x)                                                                                              \
 	({                                                                                                             \
-		uint32_t val = 0U;                                                                                     \
+		uint64_t val = 0UL;                                                                                    \
 		asm volatile("csrrc %0, sstatus, %1\n" : "=r"(val) : "i"(SSTATUS_SIE) : "memory");                     \
-		*(uint32_t *)(x) = val;                                                                                \
+		*(uint64_t *)(x) = val;                                                                                \
 	})
 
 #define CPU_INT_ALL_DISABLE(x) local_irq_save(x)
